@@ -1,11 +1,10 @@
 package ru.job4j.collection.linkedlist;
 
-import org.hamcrest.core.Is;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.Iterator;
 
-import static org.junit.Assert.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class SimpleLinkedListTest {
 
@@ -14,16 +13,18 @@ public class SimpleLinkedListTest {
         List<Integer> list = new SimpleLinkedList<>();
         list.add(1);
         list.add(2);
-        assertThat(list.get(0), Is.is(1));
-        assertThat(list.get(1), Is.is(2));
+        assertThat(list.get(0).equals(1));
+        assertThat(list.get(1).equals(1));
     }
 
-    @Test(expected = IndexOutOfBoundsException.class)
+    @Test
     public void whenGetFromOutOfBoundThenExceptionThrown() {
         List<Integer> list = new SimpleLinkedList<>();
         list.add(1);
         list.add(2);
-        list.get(2);
+        org.junit.jupiter.api.Assertions
+                .assertThrows(IndexOutOfBoundsException.class, () -> list.get(2));
+
     }
 
     @Test
@@ -31,7 +32,7 @@ public class SimpleLinkedListTest {
         List<Integer> list = new SimpleLinkedList<>();
         list.add(1);
         Iterator<Integer> it = list.iterator();
-        assertThat(it.hasNext(), Is.is(true));
+        assertThat(it.hasNext()).isTrue();
     }
 
     @Test
@@ -39,14 +40,14 @@ public class SimpleLinkedListTest {
         List<Integer> list = new SimpleLinkedList<>();
         list.add(1);
         Iterator<Integer> it = list.iterator();
-        assertThat(it.next(), Is.is(1));
+        assertThat(it.next()).isEqualTo(1);
     }
 
     @Test
     public void whenEmptyIterHashNextFalse() {
         List<Integer> list = new SimpleLinkedList<>();
         Iterator<Integer> it = list.iterator();
-        assertThat(it.hasNext(), Is.is(false));
+        assertThat(it.hasNext()).isFalse();
     }
 
     @Test
@@ -54,8 +55,8 @@ public class SimpleLinkedListTest {
         List<Integer> list = new SimpleLinkedList<>();
         list.add(1);
         Iterator<Integer> it = list.iterator();
-        assertThat(it.hasNext(), Is.is(true));
-        assertThat(it.hasNext(), Is.is(true));
+        assertThat(it.hasNext()).isTrue();
+        assertThat(it.hasNext()).isTrue();
     }
 
     @Test
@@ -64,8 +65,8 @@ public class SimpleLinkedListTest {
         list.add(1);
         list.add(2);
         Iterator<Integer> it = list.iterator();
-        assertThat(it.next(), Is.is(1));
-        assertThat(it.next(), Is.is(2));
+        assertThat(it.next()).isEqualTo(1);
+        assertThat(it.next()).isEqualTo(2);
     }
 
     @Test
@@ -74,16 +75,16 @@ public class SimpleLinkedListTest {
         list.add(1);
         list.add(2);
         Iterator<Integer> first = list.iterator();
-        assertThat(first.hasNext(), Is.is(true));
-        assertThat(first.next(), Is.is(1));
-        assertThat(first.hasNext(), Is.is(true));
-        assertThat(first.next(), Is.is(2));
-        assertThat(first.hasNext(), Is.is(false));
+        assertThat(first.hasNext()).isTrue();
+        assertThat(first.next()).isEqualTo(1);
+        assertThat(first.hasNext()).isTrue();
+        assertThat(first.next()).isEqualTo(2);
+        assertThat(first.hasNext()).isFalse();
         Iterator<Integer> second = list.iterator();
-        assertThat(second.hasNext(), Is.is(true));
-        assertThat(second.next(), Is.is(1));
-        assertThat(second.hasNext(), Is.is(true));
-        assertThat(second.next(), Is.is(2));
-        assertThat(second.hasNext(), Is.is(false));
+        assertThat(second.hasNext()).isTrue();
+        assertThat(second.next()).isEqualTo(1);
+        assertThat(second.hasNext()).isTrue();
+        assertThat(second.next()).isEqualTo(2);
+        assertThat(second.hasNext()).isFalse();
     }
 }
