@@ -1,39 +1,41 @@
 package ru.job4j.it;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.NoSuchElementException;
 
-public class BackwardArrayItTest {
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
+public class BackwardArrayItTest {
 
     @Test
     public void whenMultiCallhasNextThenTrue() {
         BackwardArrayIt it = new BackwardArrayIt(
-                new int[] {1, 2, 3}
+                new int[]{1, 2, 3}
         );
-        assertThat(it.hasNext(), is(true));
-        assertThat(it.hasNext(), is(true));
+        assertThat(it.hasNext()).isTrue();
+        assertThat(it.hasNext()).isTrue();
     }
 
     @Test
     public void whenReadSequence() {
         BackwardArrayIt it = new BackwardArrayIt(
-                new int[] {1, 2, 3}
+                new int[]{1, 2, 3}
         );
-        assertThat(it.next(), is(3));
-        assertThat(it.next(), is(2));
-        assertThat(it.next(), is(1));
+        assertThat(it.next()).isEqualTo(3);
+        assertThat(it.next()).isEqualTo(2);
+        assertThat(it.next()).isEqualTo(1);
     }
 
-    @Test(expected = NoSuchElementException.class)
+    @Test()
     public void whenNextFromEmpty() {
-        BackwardArrayIt it = new BackwardArrayIt(
-                new int[] {}
-        );
-        it.next();
+        assertThrows(NoSuchElementException.class,
+                () -> {
+                    BackwardArrayIt it = new BackwardArrayIt(
+                            new int[]{}
+                    );
+                    it.next();
+                });
     }
 }
