@@ -1,12 +1,13 @@
 package ru.job4j.iterator;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.*;
+import org.junit.jupiter.api.Test;
 
-import org.junit.Test;
 import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class FlatMapTest {
     @Test
@@ -16,9 +17,9 @@ public class FlatMapTest {
                 List.of(2, 3).iterator()
         ).iterator();
         FlatMap<Integer> flat = new FlatMap<>(data);
-        assertThat(flat.next(), is(1));
-        assertThat(flat.next(), is(2));
-        assertThat(flat.next(), is(3));
+        assertThat(flat.next()).isEqualTo(1);
+        assertThat(flat.next()).isEqualTo(2);
+        assertThat(flat.next()).isEqualTo(3);
     }
 
     @Test
@@ -27,9 +28,9 @@ public class FlatMapTest {
                 List.of(1, 2, 3).iterator()
         ).iterator();
         FlatMap<Integer> flat = new FlatMap<>(data);
-        assertThat(flat.next(), is(1));
-        assertThat(flat.next(), is(2));
-        assertThat(flat.next(), is(3));
+        assertThat(flat.next()).isEqualTo(1);
+        assertThat(flat.next()).isEqualTo(2);
+        assertThat(flat.next()).isEqualTo(3);
     }
 
     @Test
@@ -38,8 +39,8 @@ public class FlatMapTest {
                 List.of(1).iterator()
         ).iterator();
         FlatMap<Integer> flat = new FlatMap<>(data);
-        assertThat(flat.hasNext(), is(true));
-        assertThat(flat.hasNext(), is(true));
+        assertThat(flat.hasNext()).isTrue();
+        assertThat(flat.hasNext()).isTrue();
     }
 
     @Test
@@ -48,17 +49,20 @@ public class FlatMapTest {
                 List.of(1).iterator()
         ).iterator();
         FlatMap<Integer> flat = new FlatMap<>(data);
-        assertThat(flat.next(), is(1));
-        assertThat(flat.hasNext(), is(false));
+        assertThat(flat.next()).isEqualTo(1);
+        assertThat(flat.hasNext()).isFalse();
     }
 
-    @Test(expected = NoSuchElementException.class)
+    @Test
     public void whenEmpty() {
-        Iterator<Iterator<Object>> data = List.of(
-                List.of().iterator()
-        ).iterator();
-        FlatMap<Object> flat = new FlatMap<>(data);
-        flat.next();
+        assertThrows(NoSuchElementException.class,
+                () -> {
+                    Iterator<Iterator<Object>> data = List.of(
+                            List.of().iterator()
+                    ).iterator();
+                    FlatMap<Object> flat = new FlatMap<>(data);
+                    flat.next();
+                });
     }
 
     @Test
@@ -70,8 +74,8 @@ public class FlatMapTest {
                 List.of(1).iterator()
         ).iterator();
         FlatMap flat = new FlatMap(it);
-        assertTrue(flat.hasNext());
-        assertThat(1, is(flat.next()));
+        assertThat(flat.hasNext()).isTrue();
+        assertThat(1).isEqualTo(flat.next());
     }
 
     @Test
@@ -83,6 +87,6 @@ public class FlatMapTest {
                 List.of().iterator()
         ).iterator();
         FlatMap flat = new FlatMap(it);
-        assertFalse(flat.hasNext());
+        assertThat(flat.hasNext()).isFalse();
     }
 }
